Revised docker setup
The application is not any more attached to the docker image being built
but is fetched when the docker instance starts at first time.

This might be changed back in the future, but for now this allows
for easier deployment of different versions. Version can now be
specified when issuing `docker run`.

Also the configuration file is linked in as a volume instead as
attached to the docker image. This allows for easier environment
configuration.

See new README.md in Docker folder